+++ laktory/_parsers.py
@@ -1,6 +1,6 @@
 def _snake_to_camel(snake_str):
-    components = snake_str.split('_')
-    return components[0] + ''.join(x.title() for x in components[1:])
+    components = snake_str.split("_")
+    return components[0] + "".join(x.title() for x in components[1:])
 
 
 def camelize_keys(d):