+++ laktory/_testing/stockprices.py
@@ -65,7 +65,7 @@ class EventsManager:
                     )
                 ]
                 if events_root:
-                    self.events[-1].events_root = events_root
+                    self.events[-1].events_root_ = events_root
 
         return self.events
 
@@ -177,8 +177,12 @@ table_slv = Table(
     },
     expectations=[
         {"name": "positive_price", "expression": "open > 0", "action": "FAIL"},
-        {"name": "recent_price", "expression": "created_at > '2023-01-01'", "action": "DROP"},
-    ]
+        {
+            "name": "recent_price",
+            "expression": "created_at > '2023-01-01'",
+            "action": "DROP",
+        },
+    ],
 )