+++ laktory/models/databricks/notebook.py
@@ -3,9 +3,10 @@ from typing import Any
 from typing import Literal
 from typing import Union
 from pydantic import model_validator
+from laktory import constants
 from laktory.models.basemodel import BaseModel
 from laktory.models.resources.pulumiresource import PulumiResource
-from laktory.models.databricks.permission import Permission
+from laktory.models.databricks.accesscontrol import AccessControl
 from laktory.models.databricks.permissions import Permissions
 
 
@@ -15,17 +16,17 @@ class Notebook(BaseModel, PulumiResource):
 
     Attributes
     ----------
-    source:
-        Path to notebook in source code format on local filesystem.
+    access_controls:
+        List of notebook access controls
     dirpath:
         Workspace directory containing the notebook. Filename will be assumed to be the same as local filepath. Used
         if path is not specified.
-    path:
-        Workspace filepath for the notebook
     language:
          Notebook programming language
-    permissions:
-        List of notebook permissions
+    path:
+        Workspace filepath for the notebook
+    source:
+        Path to notebook in source code format on local filesystem.
 
     Examples
     --------
@@ -36,7 +37,7 @@ class Notebook(BaseModel, PulumiResource):
         source="./notebooks/pipelines/dlt_brz_template.py",
     )
     print(notebook.path)
-    #> /pipelines/dlt_brz_template.py
+    #> /.laktory/pipelines/dlt_brz_template.py
 
     notebook = models.Notebook(source="./notebooks/create_view.py", dirpath="/views/")
     print(notebook.path)
@@ -44,11 +45,11 @@ class Notebook(BaseModel, PulumiResource):
     ```
     """
 
-    source: str
+    access_controls: list[AccessControl] = []
     dirpath: str = None
-    path: str = None
     language: Literal["SCALA", "PYTHON", "SQL", "R"] = None
-    permissions: list[Permission] = []
+    path: str = None
+    source: str
 
     @property
     def filename(self) -> str:
@@ -62,7 +63,10 @@ class Notebook(BaseModel, PulumiResource):
                 self.path = f"{self.dirpath}{self.filename}"
 
             elif "/notebooks/" in self.source:
-                self.path = "/" + self.source.split("/notebooks/")[-1]
+                self.path = (
+                    constants.LAKTORY_WORKSPACE_ROOT
+                    + self.source.split("/notebooks/")[-1]
+                )
 
         return self
 
@@ -84,18 +88,16 @@ class Notebook(BaseModel, PulumiResource):
 
     @property
     def resources(self) -> list[PulumiResource]:
-
         if self.resources_ is None:
             self.resources_ = [
                 self,
             ]
-            if self.permissions:
-
+            if self.access_controls:
                 self.resources_ += [
                     Permissions(
                         resource_name=f"permissions-{self.resource_name}",
-                        access_controls=self.permissions,
-                        notebook_id=f"${{resources.{self.resource_name}.id}}",
+                        access_controls=self.access_controls,
+                        notebook_path=f"${{resources.{self.resource_name}.path}}",
                     )
                 ]
 
@@ -112,8 +114,9 @@ class Notebook(BaseModel, PulumiResource):
     @property
     def pulumi_cls(self):
         import pulumi_databricks as databricks
+
         return databricks.Notebook
 
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
-        return ["permissions", "dirpath"]
+        return ["access_controls", "dirpath"]