+++ laktory/models/databricks/permissions.py
@@ -1,10 +1,10 @@
 from laktory.models.basemodel import BaseModel
-from laktory.models.databricks.permission import Permission
+from laktory.models.databricks.accesscontrol import AccessControl
 from laktory.models.resources.pulumiresource import PulumiResource
 
 
 class Permissions(BaseModel, PulumiResource):
-    access_controls: list[Permission]
+    access_controls: list[AccessControl]
     pipeline_id: str = None
     job_id: str = None
     cluster_id: str = None
@@ -12,6 +12,7 @@ class Permissions(BaseModel, PulumiResource):
     directory_path: str = None
     experiment_id: str = None
     notebook_id: str = None
+    notebook_path: str = None
     object_type: str = None
     registered_model_id: str = None
     repo_id: str = None
@@ -35,4 +36,5 @@ class Permissions(BaseModel, PulumiResource):
     @property
     def pulumi_cls(self):
         import pulumi_databricks as databricks
+
         return databricks.Permissions