+++ laktory/models/databricks/secret.py
@@ -35,4 +35,9 @@ class Secret(BaseModel, PulumiResource):
     @property
     def pulumi_cls(self):
         import pulumi_databricks as databricks
+
         return databricks.Secret
+
+    @property
+    def pulumi_renames(self) -> dict[str, str]:
+        return {"value": "string_value"}