+++ laktory/models/databricks/workspacefile.py
@@ -2,9 +2,10 @@ import os
 from typing import Any
 from typing import Union
 from pydantic import model_validator
+from laktory import constants
 from laktory.models.basemodel import BaseModel
 from laktory.models.resources.pulumiresource import PulumiResource
-from laktory.models.databricks.permission import Permission
+from laktory.models.databricks.accesscontrol import AccessControl
 from laktory.models.databricks.permissions import Permissions
 
 
@@ -14,21 +15,21 @@ class WorkspaceFile(BaseModel, PulumiResource):
 
     Attributes
     ----------
-    source:
-        Path to file on local filesystem.
+    access_controls:
+        List of file access controls
     dirpath:
         Workspace directory containing the file. Filename will be assumed to be the same as local filepath. Used if path
         is not specified.
     path:
          Workspace filepath for the file
-    permissions:
-        List of file permissions
+    source:
+        Path to file on local filesystem.
     """
 
-    source: str
+    access_controls: list[AccessControl] = []
     dirpath: str = None
     path: str = None
-    permissions: list[Permission] = []
+    source: str
 
     @property
     def filename(self) -> str:
@@ -42,7 +43,10 @@ class WorkspaceFile(BaseModel, PulumiResource):
                 self.path = f"{self.dirpath}{self.filename}"
 
             elif "/workspacefiles/" in self.source:
-                self.path = "/" + self.source.split("/workspacefiles/")[-1]
+                self.path = (
+                    constants.LAKTORY_WORKSPACE_ROOT
+                    + self.source.split("/workspacefiles/")[-1]
+                )
 
             else:
                 raise ValueError(
@@ -70,20 +74,19 @@ class WorkspaceFile(BaseModel, PulumiResource):
 
     @property
     def resources(self) -> list[PulumiResource]:
-
         if self.resources_ is None:
-
             self.resources_ = [
                 self,
             ]
-            if self.permissions:
-
+            if self.access_controls:
                 self.resources_ += [
                     Permissions(
                         resource_name=f"permissions-{self.resource_name}",
-                        access_controls=self.permissions,
+                        access_controls=self.access_controls,
                         workspace_file_path=self.path,
-                        options={"depends_on": [f"${{resources.{self.resource_name}}}"]},
+                        options={
+                            "depends_on": [f"${{resources.{self.resource_name}}}"]
+                        },
                     )
                 ]
 
@@ -100,8 +103,9 @@ class WorkspaceFile(BaseModel, PulumiResource):
     @property
     def pulumi_cls(self):
         import pulumi_databricks as databricks
+
         return databricks.WorkspaceFile
 
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
-        return ["permissions", "dirpath"]
+        return ["access_controls", "dirpath"]