+++ laktory/models/sql/volume.py
@@ -106,12 +106,8 @@ class Volume(BaseModel, PulumiResource):
 
     @property
     def resources(self) -> list[PulumiResource]:
-
         if self.resources_ is None:
-
-            self.resources_ = [
-                self
-            ]
+            self.resources_ = [self]
 
             # Volume grants
             if self.grants:
@@ -120,12 +116,12 @@ class Volume(BaseModel, PulumiResource):
                         resource_name=f"grants-{self.resource_name}",
                         volume=self.full_name,
                         grants=[
-                            {
-                                "principal": g.principal, "privileges": g.privileges
-                            }
+                            {"principal": g.principal, "privileges": g.privileges}
                             for g in self.grants
                         ],
-                        options={"depends_on": [f"${{resources.{self.resource_name}}}"]},
+                        options={
+                            "depends_on": [f"${{resources.{self.resource_name}}}"]
+                        },
                     )
                 ]
 
@@ -142,6 +138,7 @@ class Volume(BaseModel, PulumiResource):
     @property
     def pulumi_cls(self):
         import pulumi_databricks as databricks
+
         return databricks.Volume
 
     @property