+++ tests/test_dataevent.py
@@ -5,6 +5,7 @@ import pytest
 from laktory._testing import StockPriceDataEventHeader
 from laktory._testing import EventsManager
 from laktory import models
+from laktory import settings
 
 header = StockPriceDataEventHeader()
 events = EventsManager().build_events()
@@ -12,16 +13,40 @@ event = events[0]
 
 
 def test_dataeventheader():
+    print(header.model_dump())
     assert header.model_dump() == {
         "name": "stock_price",
         "description": None,
         "producer": {"name": "yahoo-finance", "description": None, "party": 1},
-        "events_root": "/Volumes/dev/sources/landing/events/",
+        "events_root_": None,
     }
+    assert header.events_root == "/Volumes/dev/sources/landing/events/"
     assert (
         header.event_root
         == "/Volumes/dev/sources/landing/events/yahoo-finance/stock_price/"
     )
+
+
+def test_dataeventheader_root():
+    # Custom value for events root
+    header = models.DataEventHeader(name="my-event", events_root="test")
+    header2 = models.DataEventHeader(
+        **header.model_dump(by_alias=True, exclude_none=True)
+    )
+    assert header.events_root == "test"
+    assert header2.events_root == "test"
+
+    # Default value
+    header = models.DataEventHeader(name="my-event")
+    assert header.events_root == "/Volumes/dev/sources/landing/events/"
+    root0 = settings.workspace_landing_root
+    header2 = models.DataEventHeader(
+        **header.model_dump(by_alias=True, exclude_none=True)
+    )
+    settings.workspace_landing_root = "custom_root/"
+    assert header.events_root == "custom_root/events/"
+    assert header2.events_root == "custom_root/events/"
+    settings.workspace_landing_root = root0
 
 
 def test_dataevent():
@@ -40,12 +65,11 @@ def test_dataevent():
 def test_model_dump():
     # Without exclusions
     d = event.model_dump(exclude=[])
-    print(d)
     assert d == {
         "name": "stock_price",
         "description": None,
         "producer": {"name": "yahoo-finance", "description": None, "party": 1},
-        "events_root": "/Volumes/dev/sources/landing/events/",
+        "events_root": None,
         "data": {
             "created_at": "2023-09-01T00:00:00",
             "symbol": "AAPL",
@@ -61,6 +85,7 @@ def test_model_dump():
         "tstamp_col": "created_at",
         "tstamp_in_path": True,
     }
+    assert event.events_root == "/Volumes/dev/sources/landing/events/"
 
     # With exclusions
     d = event.model_dump()
@@ -126,6 +151,7 @@ def test_to_databricks_mount():
 
 if __name__ == "__main__":
     test_dataeventheader()
+    test_dataeventheader_root()
     test_dataevent()
     test_model_dump()
     test_event_without_tstamp()